Refactor e2e helper function path
