+++ crates/e2e/tests/e2e/eth_integration.rs
@@ -1,7 +1,9 @@
 use crate::{
+    onchain_components::{
+        deploy_token_with_weth_uniswap_pool, to_wei, uniswap_pair_provider, WethPoolConfig,
+    },
     services::{
-        create_order_converter, create_orderbook_api, deploy_token_with_weth_uniswap_pool, to_wei,
-        uniswap_pair_provider, wait_for_solvable_orders, OrderbookServices, WethPoolConfig,
+        create_order_converter, create_orderbook_api, wait_for_solvable_orders, OrderbookServices,
         API_HOST,
     },
     tx,
