+++ crates/e2e/tests/e2e/limit_orders.rs
@@ -1,9 +1,14 @@
 use std::sync::Arc;
 
-use crate::services::{
-    create_order_converter, create_orderbook_api, deploy_mintable_token,
-    deploy_token_with_weth_uniswap_pool, to_wei, uniswap_pair_provider, wait_for_solvable_orders,
-    OrderbookServices, WethPoolConfig, API_HOST,
+use crate::{
+    onchain_components::{
+        deploy_mintable_token, deploy_token_with_weth_uniswap_pool, to_wei, uniswap_pair_provider,
+        WethPoolConfig,
+    },
+    services::{
+        create_order_converter, create_orderbook_api, wait_for_solvable_orders, OrderbookServices,
+        API_HOST,
+    },
 };
 use contracts::IUniswapLikeRouter;
 use ethcontract::prelude::{Account, Address, PrivateKey, U256};
