+++ crates/e2e/tests/e2e/main.rs
@@ -3,10 +3,11 @@
 // here and in this test we include all the tests we want to run.
 
 #[macro_use]
-mod services;
+mod onchain_components;
 mod deploy;
 mod eth_flow;
 mod local_node;
+mod services;
 
 // Each of the following modules contains tests.
 mod eth_integration;
